fix(quick_eval/tests): test_v079_gaps.py 타입 오류 3건
1. agent_eval(m, task_type="qa", allow_duplicate_task_ids=False)는 제거된
   파라미터로 TypeError를 검증하는 의도적 테스트 —
   # type: ignore[call-arg] 추가.

2. from agent_evaluator import _FRAMEWORK_ADAPTERS — decorators.py의
   __all__에는 있지만 agent_evaluator/__init__.py 자체의 __all__에는
   빠져 있어(re-export를 명시하는 곳은 decorators.py 자체 __all__뿐,
   최상위 패키지가 그 이름을 자신의 __all__에 다시 넣어줘야 함) private
   import로 오판됐다. __init__.py의 __all__에 '_FRAMEWORK_ADAPTERS' 추가
   (agent_evaluator._FRAMEWORK_ADAPTERS는 실제로 임포트 가능한 공개
   심볼 — 소스 수정, 실제 재노출 계약과 일치).

3. QuickEval.cached(ttl: int = 3600)는 self._response_cache[key] =
   (result, _now + ttl) 형태의 순수 float 산술로만 쓰이는데(_now =
   time.time()), 이 테스트는 TTL 만료를 빠르게 검증하려고 ttl=0.01(10ms)을
   실제로 필요로 한다. AlertEngine.window_seconds에 적용한 것과 동일한
   이유로 float으로 넓힘(quick_eval.py 소스 수정, 동작 무변화).

pyright 0 errors 확인(세 파일 모두 — __init__.py의 사전 존재하던 lazy-load
전용 __all__ 경고 14건은 이번 diff와 무관). QuickEval.cached(ttl=0.01) 및
_FRAMEWORK_ADAPTERS import 실제 호출로 스모크 테스트. 이 파일 67 passed,
전체 스위트 3601 passed.

+++ agent_evaluator/quick_eval.py
@@ -1571,7 +1571,7 @@ class QuickEval:
 
     def cached(
         self,
-        ttl: int = 3600,
+        ttl: float = 3600,
         cache_key_fn: Callable | None = None,
     ) -> Callable:
         """응답 캐싱 데코레이터를 반환한다 (E5).
